# Add callback functionality for dask-delayed dataset saving