… CHANGELOG.md
… examples/pl.yaml
… trollflow2/tests/test_trollflow2.py