@gerritholl 2022-11-09
Putting on draft as I'm trying to figure out a good way for the callback function to get the information it may need.