@gerritholl 2022-11-09
When I collect a bunch of

```
    obj = fake_scene.save_dataset("dragon_top_height", filename=fn, compute=False)
```

and pass them to my wrapper, the files are there and filled.  When I collect a bunch of

```
    obj = trollflow2.plugins.save_dataset(
            {"sargasso": fake_scene},
            {"use_tmp_file": False,
             "fname_pattern": fn,
             "area": "sargasso",
             "product": "dragon_top_height"},
            {"writer": "geotiff"},
            {}, compute=False)
```

and pass them to my wrapper, the files are there but they are empty.