@gerritholl 2022-11-09
The difference is that for the NetCDF and simple_image writers, the result of `save_dataset` is a `Delayed` object.  For geotiff, it's a list of a dask array with a RIODataset object.