@gerritholl 2022-11-10
I also tried to replace

```
obj = fake_scene.save_dataset("dragon_top_height", filename=fn,
    compute=False)
obs.append(delayed(on_done)(obj, fn))
```

with
```
obj = fake_scene.save_dataset("dragon_top_height", filename=fn,
    compute=False)
    (src, targ) = obj
    objs.append(delayed(on_done)(da.store(src, targ, compute=False), fn))
```

but the problem remains, when `on_done` is called the files still have size 0