@gerritholl 2022-11-10
When I add

```python
targ.close()
```

it results in

```
Size for /tmp/test0.tif is 551
Size for /tmp/test2.tif is 551
Size for /tmp/test1.tif is 551
Size for /tmp/test4.tif is 551
Size for /tmp/test3.tif is 551
```

which corresponds to only the header