@djhoese 2022-11-10
When you run this code that you tried:

```
obj = fake_scene.save_dataset("dragon_top_height", filename=fn, compute=False)
(src, targ) = obj
objs.append(delayed(on_done)(da.store(src, targ, compute=False), fn))
```

Can you add a print to `on_done` to print out what the argument is? Is it a Delayed object? Or is it `None`? If it is `None` then I'm extremely surprised at the size 0 files. Oh but it sounds like you needed to close the file for it to flush everything to disk, maybe?