@gerritholl 2022-11-11
The value of `result` is `(None, None)`:

```
Size for /tmp/test4.tif is 551
[None, None]
Size for /tmp/test3.tif is 551
[None, None]
Size for /tmp/test2.tif is 551
[None, None]
Size for /tmp/test0.tif is 551
[None, None]
Size for /tmp/test1.tif is 551
[None, None]
```

> Oh but it sounds like you needed to close the file for it to flush everything to disk, maybe?

Right.  Example without any pytroll components:

```python
from dask import delayed
import dask.array as da
import os.path
import os
import h5py

ds = da.arange(100).reshape(10, 10)
out = "/tmp/myfile.h5"
os.unlink(out)
f = h5py.File(out, mode='a')

dset = f.create_dataset("data", ds.shape, dtype='f8')

def on_done(result, fn):
    print("Size for", fn, "is now", os.path.getsize(fn))
    return result

x = da.store(ds, dset, compute=False)

da.compute(delayed(on_done)(x, out))
print("Size after computing for", out, "is", os.path.getsize(out))
f.close()
print("Final size for", out, "is", os.path.getsize(out))
```

which results in

```
Size for /tmp/myfile.h5 is now 96
Size after computing for /tmp/myfile.h5 is 96
Final size for /tmp/myfile.h5 is 2848
```

So the issue here appears to be that computing does not result in the file being flushed/closed.  I guess for the other writers, the file gets closed before computation is finished.

We might need to put a `.flush()` somewhere.  Not sure where, and not sure what the impacts of doing so might be.