@gerritholl 2022-11-11
I've added a callback that does a close (there's no flush exposed) and it seems to work for GeoTIFF now, at least the unit tests pass.  Will do a live test later today.