@gerritholl 2022-11-11
A quick test shows that for this case, everything seems to work as expected if I loop through `obj[1]` and close all members, but why is `obj[1]` sometimes a `RIODataset` and sometimes a tuple?