@djhoese 2022-11-11
I mentioned the reasoning for the multi-element target on slack (tiff tags using dask arrays) and that we should maybe add a wrapper object in XRImage. For the `on_done` callback, should it maybe be doing the closing? Or maybe be chained after another function that does the closing?