@gerritholl 2022-11-11
It seems to pass a system test now, with one caveat: when we use the `callback_move` plugin, the launcher complains about missing files.