@gerritholl 2022-11-11
The mover callback could use `staging_zone`, so that we can have a 1-1 mapping between temporary path and final path.  Then, in `save_datasets`, In `job['produced_files'].put(fmat_config['filename'])`, we could already put the post-moving filename such that this is where `check_datasets` will look later.