@gerritholl 2022-11-22
When I use the synchronous scheduler for debugging purposes, producing FCI over Europe (LAC4) without enhancing or resampling is *faster* with callbacks than without.  It's slightly faster when the callback does nothing, but it's *much* faster when the callback is closing files.

| What | RAM, baseline | time, baseline | RAM, close/log | time, close/log | RAM, noop | time, noop
| ------------- | ------------- | ----- | ----- | ----- | ----- | ----- |
| default scheduler | 3.4 GB | 0:48 | 3.7 GB | 1:02 | 3.3 GB | 1:13 | 
| synchronous scheduler | 2.9 GB | 2:11 | 2.8 GB | 1:22 | 2.8 GB | 2:00